requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
markdown==3.5.1
jinja2==3.1.2
pyyaml==6.0.1
sqlalchemy==2.0.23
